implement interior mutability for streamer
catcodes is now moved inside the lexer struct. We couldn't
before because the borrow checker didn't allowed to get immutable
catcodes with mutable streamer. Now that streamer is borrowed as
immutable, we don't have any problem.

+++ src/lexer/catcodes.rs
@@ -22,7 +22,7 @@ impl Catcodes {
         Self::default()
     }
 
-    pub fn set_catcode(&mut self, index: i8, value: char) {
+    pub fn set_catcode(&mut self, index: usize, value: char) {
         assert!(index > 0 && index < 16);
 
         // Check if the code is used by others catcodes.
@@ -31,7 +31,7 @@ impl Catcodes {
         }
 
         // Update the codes.
-        self.codes[index as usize].push(value);
+        self.codes[index].push(value);
     }
 
     catcode_test!(is_control_sequence, 0);
